Change Compose project name and context behavior
This renames the `XF_ENVIRONMENTS` variable to `XF_CONTEXTS`, since we
often refer to each instance as an environment. It also introduces an
`XF_INSTANCE` variable for the Compose project name, which also
simplifies setting sane defaults for the board title and cookie prefix.

A `xf::get_variable` helper has been introduced for fetching an
environment variable and falling back to the `.env` value.

A `xf::get_contexts` helper has been introduced for parsing the context
value.